+++ airbyte-cdk/python/unit_tests/sources/declarative/async_job/test_job_orchestrator.py
@@ -163,13 +163,53 @@ class AsyncJobOrchestratorTest(TestCase):
                 self._job_for_another_slice: [AsyncJobStatus.COMPLETED],
             }
         )
-
         orchestrator = AsyncJobOrchestrator(self._job_repository, [self._job_for_a_slice.job_parameters(), self._job_for_another_slice.job_parameters()], job_tracker)
 
         partitions = list(orchestrator.create_and_get_completed_partitions())
 
         assert len(partitions) == 2
         assert job_tracker.try_to_get_intent()
+
+    @mock.patch(sleep_mock_target)
+    def test_given_exception_when_start_job_and_raise_this_exception_and_abort_jobs(self, mock_sleep: MagicMock) -> None:
+        orchestrator = AsyncJobOrchestrator(
+            self._job_repository,
+        [_A_STREAM_SLICE, _ANOTHER_STREAM_SLICE],
+            JobTracker(_NO_JOB_LIMIT),
+            exceptions_to_break_on=[ValueError],
+        )
+        self._job_repository.start.side_effect = [self._job_for_a_slice, ValueError("Something went wrong")]
+
+        with pytest.raises(ValueError):
+            # assert that orchestrator exits on expected error
+            list(orchestrator.create_and_get_completed_partitions())
+        assert len(orchestrator._running_partitions) == 0
+        self._job_repository.abort.assert_called_once_with(self._job_for_a_slice)
+
+    @mock.patch(sleep_mock_target)
+    def test_given_exception_when_start_job_and_skip_this_exception(self, mock_sleep: MagicMock) -> None:
+        self._job_repository.start.side_effect = [
+            self._job_for_a_slice,
+            AirbyteTracedException("Something went wrong. Expected error"),
+            self._job_for_another_slice,
+        ]
+        self._job_repository.update_jobs_status.side_effect = _status_update_per_jobs(
+            {
+                self._job_for_a_slice: [AsyncJobStatus.COMPLETED],
+                self._job_for_another_slice: [AsyncJobStatus.RUNNING, AsyncJobStatus.COMPLETED],
+            }
+        )
+        orchestrator = AsyncJobOrchestrator(
+            self._job_repository,
+            [_A_STREAM_SLICE, _A_STREAM_SLICE,
+             _ANOTHER_STREAM_SLICE],
+            JobTracker(_NO_JOB_LIMIT),
+            exceptions_to_break_on=[ValueError],
+        )
+
+        partitions = list(orchestrator.create_and_get_completed_partitions())
+
+        assert len(partitions) == 2
 
     @mock.patch(sleep_mock_target)
     def test_given_jobs_failed_more_than_max_attempts_when_create_and_get_completed_partitions_then_free_job_budget(self, mock_sleep: MagicMock) -> None:
@@ -211,14 +251,14 @@ class AsyncJobOrchestratorTest(TestCase):
         job_tracker = JobTracker(1)
         self._job_repository.start.side_effect = ValueError("Can't create job")
 
-        orchestrator = AsyncJobOrchestrator(self._job_repository, [_A_STREAM_SLICE], job_tracker)
+        orchestrator = AsyncJobOrchestrator(self._job_repository, [_A_STREAM_SLICE], job_tracker, [ValueError])
 
         with pytest.raises(Exception):
             list(orchestrator.create_and_get_completed_partitions())
 
         assert job_tracker.try_to_get_intent()
 
-    def _mock_repository(self):
+    def _mock_repository(self) -> None:
         self._job_repository = Mock(spec=AsyncJobRepository)
 
     def _an_async_job(self, job_id: str, stream_slice: StreamSlice) -> AsyncJob: